backend/requirements.txt
fastapi
uvicorn
toml
edge-tts
PyPDF2
pytesseract
Pillow
python-multipart
